+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Revision history for byteslice
 
-## 0.2.10.0 -- 2023-??-??
+## 0.2.10.0 -- 2023-05-01
 
 * Add `equals13`, `equals14`, `equals15`.
 * Add `Show` instances for `BytesN` and `ByteArrayN`.